fix detail dv tag
