+++ packages/next-common/components/comment/polkassemblyUser.jsx
@@ -1,13 +1,18 @@
 import { useDetailType } from "next-common/context/page";
 import { detailPageCategory } from "next-common/utils/consts/business/category";
-import PolkassemblyUserWithDVTag from "../user/polkassemblyUserWithDV";
 import PolkassemblyUser from "next-common/components/user/polkassemblyUser";
+import DVTag from "../detail/common/DVTag";
 
 export default function PolkassemblyCommentUser({ user }) {
   const type = useDetailType();
 
   if (type === detailPageCategory.GOV2_REFERENDUM) {
-    return <PolkassemblyUserWithDVTag user={user} />;
+    return (
+      <div className="flex items-center gap-x-2">
+        <PolkassemblyUser user={user} />
+        <DVTag address={user.address} />
+      </div>
+    );
   }
 
   return <PolkassemblyUser user={user} />;
