+++ packages/next-common/components/comment/user.jsx
@@ -1,13 +1,18 @@
 import { useDetailType } from "next-common/context/page";
 import { detailPageCategory } from "next-common/utils/consts/business/category";
+import DVTag from "../detail/common/DVTag";
 import SystemUser from "../user/systemUser";
-import SystemUserWithDVTag from "../user/systemUserWithDV";
 
 export default function CommentUser({ author }) {
   const type = useDetailType();
 
   if (type === detailPageCategory.GOV2_REFERENDUM) {
-    return <SystemUserWithDVTag user={author} />;
+    return (
+      <div className="flex items-center gap-x-2">
+        <SystemUser user={author} />
+        <DVTag address={author.address} />
+      </div>
+    );
   }
 
   return <SystemUser user={author} />;
